requirements.txt
# Core Data Analysis
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Social Media APIs
praw==7.7.1  # Reddit API
tweepy==4.14.0  # Twitter API

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Data Visualization
plotly==5.18.0
wordcloud==1.9.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Date handling
python-dateutil==2.8.2
